backend/requirements.txt
# Flask Framework
Flask==3.1.0
Flask-CORS==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.4.1
python-socketio==5.11.4

# Database
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36

# Firebase Authentication
firebase-admin==6.6.0

# Telegram Bot
python-telegram-bot==21.9

# WebSocket client for Telegram bot
python-socketio[client]==5.11.4

# PPTX Manipulation
python-pptx==1.0.2

# PDF and Image Conversion
pdf2image==1.17.0
Pillow==10.4.0

# Environment Variables
python-dotenv==1.0.1

# Utilities
requests==2.32.3

# Google Gemini AI
google-generativeai==0.8.5

# Production Server
gunicorn==21.2.0

# Redis for SocketIO message queue (multi-worker support)
redis==5.0.1
kombu==5.4.2  # Required for SocketIO Redis adapter
